Refatorado o kanban para um melhor clean code diminuindo suas quantidades de linha criando mais componentes contendo handlers e partes do visual do kanban como header,form e dialog em componentes divididos cada um com seu componente alem da utilização de wrappers para os handles importados no kanban.
