+++ frontend/src/dashboard/components/components-kanban/DialogDelete.tsx
